actualizando diseño y editar datos enviados

+++ frontend/src/pages/Usuario.jsx
@@ -1,45 +1,52 @@
 import React, { useState, useEffect } from "react";
-import { useNavigate } from "react-router-dom";
+import { useNavigate, useLocation, useSearchParams } from "react-router-dom";
 import axios from "./../components/axiosConfig";
 import "../assets/Usuario.css";
-// Datos simulados
 
-export default function Usuario(usuario) {
+export default function Usuario({
+  permisos,
+  usuario,
+  setMessage,
+  setTypeMessage,
+}) {
   const [usuarios, setUsuarios] = useState([]);
   const [usuarioSeleccionado, setUsuarioSeleccionado] = useState(null);
   const [modoEdicion, setModoEdicion] = useState(false);
   const [formData, setFormData] = useState({ nombre: "", rol: "" });
   const [registros, setRegistros] = useState([]);
-  const [permisos, setPermisos] = useState([]);
+
+  const [rol, setRol] = useState([]);
+  const [registrar, setRegistrar] = useState("");
 
   const [abierto, setAbierto] = useState(false);
   const [seleccionado, setSeleccionado] = useState(null);
 
-  const navigate = useNavigate();
-  // Cargar imágenes desde el servidor
-  useEffect(() => {
-    listaUsuarios();
-    //setUsuarios(listaCategory.data);
-    obtenerPermisos();
-  }, [navigate]);
+  const [anterior, setAnterior] = useState(true);
+  const [siguiente, setSiguiente] = useState(true);
 
-  const obtenerPermisos = async () => {
-    try {
-      const response = await axios.get("/api/permisos/usuario");
-      setPermisos(response.data);
-    } catch (error) {
-      console.log(error);
-    }
-  };
+  const [searchParams, setSearchParams] = useSearchParams();
+  const [cantidad] = useState(parseInt(searchParams.get("cantidad")) || 6);
+  const [paginaActual, setPaginaActual] = useState(
+    parseInt(searchParams.get("paginaActual")) || 1
+  );
+
+  const params = new URLSearchParams();
+  params.append("paginaActual", paginaActual);
+  params.append("cantidad", cantidad);
+
+  const location = useLocation();
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    listaUsuarios(paginaActual);
+    listarRoles();
+  }, [navigate]);
 
   const formatearFecha = (fecha) => {
     const opciones = {
       year: "numeric",
       month: "2-digit",
       day: "2-digit",
-      // hour: "2-digit",
-      // minute: "2-digit",
-      // second: "2-digit",
       hour12: false,
     };
 
@@ -48,36 +55,50 @@ export default function Usuario(usuario) {
       .replace(",", "");
   };
 
-  const listaUsuarios = async () => {
+  const listaUsuarios = async (pagina) => {
     try {
-      const response = await axios.get("/api/usuarios");
+      const response = await axios.get("/api/usuarios", {
+        params,
+      });
+      if (location.pathname !== "/perfil/permiso/usuarios") {
+        navigate({
+          pathname: "/perfil/permiso/usuarios",
+          search: `?params=${params}`,
+        });
+      } else {
+        setSearchParams({ params });
+      }
+      setSiguiente(true);
+      setAnterior(true);
+      if (response.data.count / cantidad - pagina > 0) setSiguiente(false);
 
-      setUsuarios(response.data); // Actualizar el estado con las categorías
+      if (pagina > 1) setAnterior(false);
+
+      setUsuarios(response.data.rows);
     } catch (error) {
-      console.log(error);
+      console.error(error);
     }
   };
 
-  const rol = [
-    { id: "1", nombre: "Admin" },
-    { id: "2", nombre: "Default" },
-    { id: "3", nombre: "Categoria" },
-    { id: "5", nombre: "User" },
-  ];
+  const listarRoles = async () => {
+    try {
+      const response = await axios.get(`/api/permisos/`);
+      setRol(response.data);
+    } catch (error) {
+      console.error(error);
+    }
+  };
 
   const listarRegistros = async (id) => {
     try {
       const response = await axios.get(`/api/registros/registros/${id}`);
-
-      setRegistros(response.data); // Actualizar el estado con las categorías
+      setRegistros(response.data);
     } catch (error) {
-      console.log(error);
+      console.error(error);
     }
   };
 
   const handleSeleccionar = (usuario) => {
-    listarRegistros;
-
     setUsuarioSeleccionado(usuario);
     setModoEdicion(false);
     const valor = rol.find((ro) => ro.id == usuario.Permiso.id);
@@ -92,7 +113,23 @@ export default function Usuario(usuario) {
 
   const handleEliminar = async () => {
     if (!usuarioSeleccionado) return;
-    await axios.delete(`/api/usuarios/${usuarioSeleccionado.id}`);
+    try {
+      const confirmar = window.confirm(
+        `¿Estás seguro de que quieres eliminar el usuario "${usuarioSeleccionado.nombre}"?`
+      );
+      if (!confirmar) {
+        // Usuario canceló, salgo de la función
+        return;
+      }
+
+      await axios.delete(`/api/usuarios/${usuarioSeleccionado.id}`);
+      setMessage("Usuario eliminado");
+      setTypeMessage("exito");
+    } catch (err) {
+      setMessage("Error: " + err);
+      setTypeMessage("error");
+    }
+
     setUsuarios(usuarios.filter((u) => u.id !== usuarioSeleccionado.id));
     setUsuarioSeleccionado(null);
     setModoEdicion(false);
@@ -103,8 +140,16 @@ export default function Usuario(usuario) {
   };
 
   const handleGuardar = async () => {
-    await axios.put(`/api/usuarios/${usuarioSeleccionado.id}`, formData);
-    listaUsuarios();
+    try {
+      await axios.put(`/api/usuarios/${usuarioSeleccionado.id}`, formData);
+      setMessage("Usuario actualizado");
+      setTypeMessage("exito");
+    } catch (err) {
+      setMessage("Error: " + err);
+      setTypeMessage("error");
+    }
+
+    listaUsuarios(paginaActual);
     setRegistros([]);
     setFormData({ nombre: "", rol: "" });
     setUsuarioSeleccionado(null);
@@ -122,11 +167,23 @@ export default function Usuario(usuario) {
 
   const handleResetearPassword = async () => {
     if (usuarioSeleccionado) {
-      await axios.post(`/api/registros/`, usuarioSeleccionado);
-      alert(
-        `Se ha enviado un enlace para resetear la contraseña a: ${usuarioSeleccionado.email}`
-      );
-      // Aquí podrías implementar lógica real con backend o un modal
+      try {
+        const confirmar = window.confirm(
+          `¿Estás seguro de que quieres resetear la contraseña al usuario "${usuarioSeleccionado.nombre}"?`
+        );
+        if (!confirmar) {
+          // Usuario canceló, salgo de la función
+          return;
+        }
+        setMessage("Cambiando contraseña...");
+        setTypeMessage("exito");
+        await axios.post(`/api/email/update`, usuarioSeleccionado);
+        setMessage("Contraseña reseteada, email enviado");
+        setTypeMessage("exito");
+      } catch (err) {
+        setMessage("Error: " + err);
+        setTypeMessage("error");
+      }
     }
   };
 
@@ -136,11 +193,50 @@ export default function Usuario(usuario) {
     setAbierto(false);
   };
 
+  const handleSubmit = async (e) => {
+    e.preventDefault();
+    try {
+      setMessage("Registrando usuario...");
+      setTypeMessage("exito");
+      await axios.post("/api/email/", { email: registrar });
+      setMessage("Usuario agregado");
+      setTypeMessage("exito");
+      listaUsuarios(paginaActual);
+    } catch (error) {
+      console.error("Error al agregar usuario:", error);
+      setMessage("Error: " + error);
+      setTypeMessage("error");
+    }
+    setRegistrar("");
+  };
+
+  const cambiarpagina = async (pagina) => {
+    params.delete("paginaActual");
+    params.append("paginaActual", pagina);
+
+    listaUsuarios(pagina);
+  };
+
   return (
     <>
       <div className="contenedor">
         {/* IZQUIERDA: info usuario + categorías */}
         <div className="columna-izquierda">
+          {permisos?.agusuario === true && (
+            <div className="usuario-register">
+              <h2>Nuevo usuario</h2>
+              <form onSubmit={handleSubmit}>
+                <input
+                  type="email"
+                  value={registrar}
+                  placeholder="Ingrese email del usuario"
+                  onChange={(e) => setRegistrar(e.target.value)}
+                />
+                <button type="submit">Registar Usuario</button>
+              </form>
+            </div>
+          )}
+          <br />
           <div className="usuario-info">
             <h2>Información del Usuario</h2>
             {usuarioSeleccionado ? (
@@ -181,13 +277,6 @@ export default function Usuario(usuario) {
                         ))}
                       </ul>
                     )}
-
-                    {/* Hidden input para enviar valor si es parte de un form */}
-                    <input
-                      type="hidden"
-                      name="categoria"
-                      value={seleccionado?.id || ""}
-                    />
                   </div>
 
                   <div className="usuario-actions">
@@ -206,9 +295,7 @@ export default function Usuario(usuario) {
                   <p>
                     <strong>Rol:</strong> {usuarioSeleccionado.Permiso.nombre}
                   </p>
-                  <p>
-                    <strong>Estado:</strong> Activo
-                  </p>
+
                   <div className="usuario-actions">
                     {permisos?.edusuario === true && (
                       <button onClick={handleEditar}>Editar</button>
@@ -218,16 +305,24 @@ export default function Usuario(usuario) {
                         {usuarioSeleccionado.id === usuario.usuario ? (
                           <button disabled>Eliminar</button>
                         ) : (
-                          <button onClick={handleEliminar}>Eliminar</button>
+                          <button onClick={() => handleEliminar()}>
+                            Eliminar
+                          </button>
                         )}
                       </>
                     )}
                     {permisos?.edusuario === true && (
-                      <button onClick={handleResetearPassword}>
+                      <button onClick={() => handleResetearPassword()}>
                         Resetear contraseña
                       </button>
                     )}
                   </div>
+                  {permisos?.elusuario === true && (
+                    <p>
+                      Eliminar: TAMBIEN ELIMINA LOS ARCHIVOS SUBIDOS POR EL
+                      USUARIO
+                    </p>
+                  )}
                 </>
               )
             ) : (
@@ -267,6 +362,29 @@ export default function Usuario(usuario) {
         {permisos?.verusuario === true && (
           <div className="columna-derecha">
             <h3>Lista de Usuarios</h3>
+            {!(siguiente && anterior) && (
+              <div className="div-paginar">
+                <button
+                  onClick={() => {
+                    cambiarpagina(paginaActual - 1);
+                    setPaginaActual(paginaActual - 1);
+                  }}
+                  disabled={anterior}
+                >
+                  Anterior
+                </button>
+                <p>{paginaActual}</p>
+                <button
+                  onClick={() => {
+                    cambiarpagina(paginaActual + 1);
+                    setPaginaActual(paginaActual + 1);
+                  }}
+                  disabled={siguiente}
+                >
+                  Siguiente
+                </button>
+              </div>
+            )}
             <br />
             {usuarios.map((u) => (
               <div key={u.id} className="usuario-item">
